Improved dupe fixing

+++ pages/api/v1/addConfig.ts
@@ -75,29 +75,43 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
             }
         })
 
+        const amount = await prisma.config.count({
+            where: {
+                name: {
+                    equals: name
+                },
+                plugin: {
+                    equals: plugin,
+                    mode: 'insensitive'
+                }
+            }
+        })
+
         if (existing != null) {
             res.status(400).json({
                 message: `Config with identical name already exists for plugin!`
             })
 
             // Perform cleanup on the database
-            await prisma.config.deleteMany({
-                where: {
-                    name: {
-                        equals: name
-                    },
-                    plugin: {
-                        equals: plugin,
-                        mode: 'insensitive'
+            if (amount > 1) {
+                await prisma.config.deleteMany({
+                    where: {
+                        name: {
+                            equals: name
+                        },
+                        plugin: {
+                            equals: plugin,
+                            mode: 'insensitive'
+                        }
                     }
-                }
-            })
+                })
 
-            await prisma.config.create({
-                data: {
-                    ...existing
-                }
-            })
+                await prisma.config.create({
+                    data: {
+                        ...existing
+                    }
+                })
+            }
 
             return
         }
